requirements.txt
pybluez
pyscreenshot
pillow